refactor generic puppet
