--- libs/media/src/lib/file/puppets/puppets.component.html
@@ -1,33 +1,0 @@
-
-<section fxLayout="column" fxLayoutAlign="space-between" [ngClass]="{ owner: event.isOwner }">
-  <ng-container *ngIf="!!event.meta.selectedFile">
-    <ng-container [ngSwitch]="event.meta.selectedFile | fileType">
-      <ng-container *ngSwitchCase="'image'">
-        <img [ref]="event.meta.selectedFile" asset="no_titles.webp" />
-      </ng-container>
-      <ng-container *ngSwitchCase="'pdf'">
-        <pdf-viewer [eventId]="event.id" [ref]="event.meta.selectedFile" [control]="event.meta.controls[event.meta.selectedFile]"></pdf-viewer>
-      </ng-container>
-      <ng-container *ngSwitchCase="'video'">
-        <video-viewer [eventId]="event.id" [ref]="event.meta.selectedFile" [control]="event.meta.controls[event.meta.selectedFile]"></video-viewer>
-      </ng-container>
-      <ng-container *ngSwitchDefault>
-        <article>
-          <h3>Sorry</h3>
-          <img asset="no_titles.webp" alt="Flying saucer">
-          <p>We don't know how to display this file.</p>
-          <p class="mat-caption">Supported files are: pdfs, images, and videos</p>
-        </article>
-      </ng-container>
-    </ng-container>
-  </ng-container>
-  <ng-container *ngIf="event.isOwner">
-    <article fxLayout="column" fxLayoutAlign="center center">
-      <mat-divider></mat-divider>
-      <button mat-button (click)="stop()">
-        <mat-icon svgIcon="cancel"></mat-icon>
-        <span>Stop Sharing</span>
-      </button>
-    </article>
-  </ng-container>
-</section>
