+++ libs/media/src/lib/file/viewers/viewers.component.scss
@@ -1,0 +1,7 @@
+
+
+img {
+  height: 100%;
+  width: 100%;
+  object-fit: contain;
+}
